requirements.txt
asgiref==3.2.3
Django==2.1
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-mptt==0.11.0
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.1
